requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
psycopg2-binary>=2.9.9

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Machine Learning
torch>=2.1.0
transformers>=4.36.0
pillow>=10.1.0

# Testing
pytest>=7.4.0
httpx>=0.25.0

# Development
python-dotenv>=1.0.0
